Refactor "FAQCategories" dataloader to class to wrap inner logic
